backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
supabase==2.13.0
resend==2.7.0
python-multipart==0.0.20
